Add options saving/loading/modifying
Requires storage permission. Options are stored as they are displayed in
the options page, and loaded based on the object types in the default
values.

All the main code is wrapped in the loadOptions.then() to be sure that
we load options from storage before anything else.

A message is sent to the background script to trigger a new loadOptions()
whenever new values are saved, so that it stays up to date.